build 16 - misc

+++ paintball2/qcommon/common.c
@@ -1397,14 +1397,14 @@ byte	COM_BlockSequenceCRCByte (byte *base, int length, int sequence)
 
 //========================================================
 
-float	frand(void)
+float frand (void)
 {
-	return (rand()&32767)* (1.0/32767);
+	return (float)(rand() & 32767) * (1.0f / 32767.0f);
 }
 
-float	crand(void)
+float crand (void)
 {
-	return (rand()&32767)* (2.0/32767) - 1;
+	return (float)(rand() & 32767) * (2.0f / 32767.0f) - 1.0f;
 }
 
 void Key_Init (void);
